+++ modules/School Admin/rollGroup_manage.php
@@ -20,7 +20,7 @@ along with this program. If not, see <http://www.gnu.org/licenses/>.
 use Gibbon\Tables\DataTable;
 use Gibbon\Services\Format;
 use Gibbon\Domain\School\SchoolYearGateway;
-use Gibbon\Domain\School\RollGroupGateway;
+use Gibbon\Domain\RollGroups\RollGroupGateway;
 
 if (isActionAccessible($guid, $connection2, '/modules/School Admin/rollGroup_manage.php') == false) {
     //Acess denied
